+++ framework/packages/standards/dex/src/msg.rs
@@ -139,8 +139,8 @@ pub enum DexQueryMsg {
     GenerateMessages {
         /// Execute message to generate messages for
         message: DexExecuteMsg,
-        /// Proxy Addr generate messages for
-        proxy_addr: String,
+        /// Sender Addr generate messages for
+        addr_as_sender: String,
     },
     /// Fee info for using the different dex actions
     #[returns(DexFeesResponse)]