+++ integrations/bundles/wyndex/tests/abstract.rs
@@ -1,12 +1,8 @@
-use cosmwasm_std::Addr;
-
 mod abstrct {
     use abstract_interface::Abstract;
-    use cosmwasm_std::Empty;
-    use cw_orch::{deploy::Deploy, mock::Mock};
+    use cosmwasm_std::{Addr, Empty};
+    use cw_orch::prelude::*;
     use wyndex_bundle::{WynDex, WYNDEX_OWNER};
-
-    use super::*;
 
     #[test]
     fn deploy() {