+++ framework/CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
   
 - Added a `.execute` method on the AuthZ API to execute `CosmosMsg` types on behalf of a granter.
+- Abstract Client builder: register dexes on ANS
+- `.sub_accounts` method on `Account` for getting Abstract Client Sub Accounts
+- Publish adapter method of Abstract Client Publisher now returns Adapter object
+- Added a `.account_from` method on the `AbstractClient` for retrieving `Account`s.
+- Creating Sub Account from `AbstractClient` Account builder.
 
 ### Changed
 
@@ -18,6 +23,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+
+- Creating sub-account from account factory is restricted. Use Create Sub Account method of the manager instead
 
 ## [0.20.0] - 2024-01-24